refactor(engine): split context.rs tests into context/tests.rs
context.rs had grown to ~950 lines, over half of it test code. Move the unit
tests, the hardware smoke tests, and the shared round-trip helpers into
context/tests.rs, leaving context.rs as just the production context. The inner
mock test module is renamed `tests` -> `mock` so paths read context::tests::mock
rather than context::tests::tests.
